Added simple and L2 regularization from TG
Hi Mira,

I played around a bit with the regularized NNLS we used. I tried to speed up the regularization parameter estimation as discussed, but in the process I realized, that the optimization is not only slow. Also the result is in many cases overregularized. Furthermore it is also bad for the transferability, that there is a different regularization for the individual patients.

So I tried to come up with a better estimate for the regularization parameter. I assumed, that the regularization term should contribute as much as the data consistency term to the overall sum of squares. With that the lambda should be on the order of [number of b-values/SNR]. This would amount to about 0.1 for your datasets.

Attached you can find a simplified version of the CVNNLS, that uses the regularization parameter lambda as the 3rd input argument. The code should be faster by a factor of 10-20.

You can now use the same lambda for all data you analyze, yielding potentially a better transferability of the results.

With the value of 0.1 the regularization should be also weaker than before allowing sharper peaks and better peak separation.

The code also outputs the completely unregularized NNLS result, which might also be helpful. In many cases the non-negativity constraint is good enough to get useful results.

This result was already computed before, but just used for comparing the residues and never saved.

I also changed the regularization slightly. The previous code always regularized down the edges of the spectrum (assuming, that there is no diffusivity outside the given range). This always annoyed me a bit, since it is slightly deceiving you. You think, that there is never a component in the spectra, that hits the ends, but this is not necessarily true. The new code now only regularizes the derivative at the edges of the spectrum. This allows peaks to hit the boarder, if that’s in the data.

In case you want to try it, I also generated a version, that adds a small L2-regularization to the curvature regularization.

Best,

Thomas